Update p2.java
double to float

+++ p2.java
@@ -2,14 +2,14 @@ import java.util.Scanner;
 
 public class p2 {
     public static void main(String[] args) {
-        double cost = 100;
+        float cost = 100;
         int unit = 1;
         Scanner s = new Scanner(System.in);
         System.out.print("Enter unit to buy :  ");
         unit = s.nextInt();
 
         if (unit > 100) {
-            cost *= 0.9;
+            cost *= (float)0.9;
         }
         System.out.print("total cost of unit is :  " + cost * unit);
